+ support action jump

+++ core/src/main/java/com/angcyo/core/component/accessibility/parse/ActionBean.kt
@@ -40,19 +40,19 @@ data class ActionBean(
     var handleType: Int = HANDLE_TYPE_NONE,
 
     /**默认当前页面检测x次, 都还不通过. 回退到上一步*/
-    var rollbackCount: Int = DEFAULT_ACTION_CHECK_OUT_MAX_COUNT,
+    var rollbackCount: Long = DEFAULT_ACTION_CHECK_OUT_MAX_COUNT,
 
     /**回滚x次后, 还是不通过, 则报错*/
-    var rollbackMaxCount: Int = DEFAULT_ROLLBACK_MAX_COUNT,
+    var rollbackMaxCount: Long = DEFAULT_ROLLBACK_MAX_COUNT,
 
     /**[checkOtherEvent]允许执行的最大次数*/
-    var checkOtherCount: Int = DEFAULT_ACTION_OTHER_MAX_COUNT,
+    var checkOtherCount: Long = DEFAULT_ACTION_OTHER_MAX_COUNT,
 
     /**允许[doAction]执行的最大次数, 超过后抛出异常*/
-    var actionMaxRunCount: Int = DEFAULT_ACTION_MAX_COUNT,
+    var actionMaxRunCount: Long = DEFAULT_ACTION_MAX_COUNT,
 
     /**当action识别到并处理执行后的次数大于此值时, 强制完成*/
-    var actionMaxCount: Int = DEFAULT_ACTION_FINISH_MAX_COUNT,
+    var actionMaxCount: Long = DEFAULT_ACTION_FINISH_MAX_COUNT,
 
     /**当前action执行完成后, 间隔多久执行下一个[Action]. 毫秒
      * 格式[5000,500,5] 解释:5000+500*[1-5),
